Don't change if a softlink already exists

+++ install.sh
@@ -17,7 +17,10 @@ function install_dot_file() {
 }
 
 install_dot_file .vimrc vimrc
-ln -sf ~/dotfiles/local/vimrc_local ~/.vimrc_local
+
+if [[ ! -L ~/.vimrc_local ]]; then
+ ln -sf ~/dotfiles/local/vimrc_local ~/.vimrc_local
+fi
 install_dot_file .vim vim
 install_dot_file .tmux.conf tmux.conf
 install_dot_file .tmux tmux
